feat(ai): cancel ACP CLI runs, replay tool transcripts to Copilot, extract message view
- ACP CLI backend (Codex/Claude) streaming sessions now honor Stop: the
  session polls the cancellation generation between received lines
  (<=250ms latency) and aborts, killing the CLI child process on drop;
  non-streaming watchdog usage is unaffected
- Copilot prompt stitching now appends the same compact per-turn tool
  transcript the OpenAI-compatible history paths replay, via a shared
  agent_tool_transcript helper
- Extract AssistantMessageView, AssistantSecretEntryCard, and the
  secret-store helpers from AssistantPanel.tsx into
  AssistantMessageView.tsx (panel 3238 -> 2867 lines); update the
  dialog-portal-policy guard to read the new file
